Extract SalesChannelType to separate file. Make compile
