+++ Modules/Sources/NetworkingCore/Model/Order.swift
@@ -1,37 +1,6 @@
 import Foundation
 import Codegen
 
-public enum SalesChannelType {
-    case pointOfSale
-}
-
-extension SalesChannelType: RawRepresentable {
-    public init?(rawValue: String) {
-        switch rawValue {
-        case "pos-rest-api":
-            self = .pointOfSale
-        default:
-            return nil
-        }
-    }
-
-    public var rawValue: String {
-        switch self {
-        case .pointOfSale:
-            return description
-        }
-    }
-
-    public var description: String {
-        switch self {
-        case .pointOfSale:
-            return NSLocalizedString("",
-                                     value: "POS",
-                                     comment: "The acronym for 'Point of Sale' that is shown for certain orders.")
-        }
-    }
-}
-
 /// Represents an Order Entity.
 ///
 public struct Order: Decodable, Sendable, GeneratedCopiable, GeneratedFakeable {
