+++ WooCommerce/WooCommerceTests/ViewRelated/Orders/OrderListCellViewModelTests.swift
@@ -70,7 +70,7 @@ final class OrderListCellViewModelTests: XCTestCase {
         let viewModel = OrderListCellViewModel(order: order, currencySettings: ServiceLocator.currencySettings)
 
         // Then
-        XCTAssertEqual(viewModel.salesChannel, OrderListCellViewModel.Localization.salesChannelPOSText)
+        XCTAssertEqual(viewModel.salesChannel, "POS")
     }
 
     func test_salesChannel_when_createdVia_is_nil_then_returns_nil() {
